quick refactor
