+++ modules/claims_api/app/models/claims_api/auto_established_claim.rb
@@ -2,12 +2,13 @@
 
 require_dependency 'claims_api/form_526'
 require_dependency 'claims_api/json_marshal'
+require_dependency 'claims_api/concerns/file_data_validation'
 
 module ClaimsApi
   class AutoEstablishedClaim < ApplicationRecord
+    include FileDataValidation
     attr_encrypted(:form_data, key: Settings.db_encryption_key, marshal: true, marshaler: ClaimsApi::JsonMarshal)
     attr_encrypted(:auth_headers, key: Settings.db_encryption_key, marshal: true, marshaler: ClaimsApi::JsonMarshal)
-    attr_encrypted(:file_data, key: Settings.db_encryption_key, marshal: true, marshaler: ClaimsApi::JsonMarshal)
 
     has_many :supporting_documents, dependent: :destroy
 
@@ -64,20 +65,6 @@ module ClaimsApi
 
     def status_from_phase(*)
       status
-    end
-
-    def file_name
-      file_data['filename']
-    end
-
-    def document_type
-      file_data['doc_type']
-    end
-
-    def set_file_data!(file_data, doc_type)
-      uploader.store!(file_data)
-      self.file_data = { filename: uploader.filename,
-                         doc_type: doc_type }
     end
 
     def uploader
