+++ modules/claims_api/app/models/claims_api/power_of_attorney.rb
@@ -1,12 +1,13 @@
 # frozen_string_literal: true
 
 require_dependency 'claims_api/json_marshal'
+require_dependency 'claims_api/concerns/file_data_validation'
 
 module ClaimsApi
   class PowerOfAttorney < ApplicationRecord
+    include FileDataValidation
     attr_encrypted(:form_data, key: Settings.db_encryption_key, marshal: true, marshaler: ClaimsApi::JsonMarshal)
     attr_encrypted(:auth_headers, key: Settings.db_encryption_key, marshal: true, marshaler: ClaimsApi::JsonMarshal)
-    attr_encrypted(:file_data, key: Settings.db_encryption_key, marshal: true, marshaler: ClaimsApi::JsonMarshal)
 
     PENDING = 'pending'
     UPDATED = 'updated'
@@ -30,20 +31,6 @@ module ClaimsApi
     def set_md5
       headers = auth_headers.except('va_eauth_issueinstant', 'Authorization')
       self.md5 = Digest::MD5.hexdigest form_data.merge(headers).to_json
-    end
-
-    def file_name
-      file_data['filename']
-    end
-
-    def document_type
-      file_data['doc_type']
-    end
-
-    def set_file_data!(file_data, doc_type)
-      uploader.store!(file_data)
-      self.file_data = { filename: uploader.filename,
-                         doc_type: doc_type }
     end
 
     def uploader
