+++ modules/claims_api/app/models/claims_api/supporting_document.rb
@@ -1,10 +1,11 @@
 # frozen_string_literal: true
 
 require_dependency 'claims_api/json_marshal'
+require_dependency 'claims_api/concerns/file_data_validation'
 
 module ClaimsApi
   class SupportingDocument < ApplicationRecord
-    attr_encrypted(:file_data, key: Settings.db_encryption_key, marshal: true, marshaler: ClaimsApi::JsonMarshal)
+    include FileDataValidation
 
     belongs_to :auto_established_claim
     validates :auto_established_claim_id, presence: true
@@ -13,25 +14,6 @@ module ClaimsApi
 
     def evss_claim_id
       auto_established_claim.evss_id
-    end
-
-    def file_name
-      file_data['filename']
-    end
-
-    def document_type
-      file_data['doc_type']
-    end
-
-    def description
-      file_data['description']
-    end
-
-    def set_file_data!(file_data, doc_type, description)
-      uploader.store!(file_data)
-      self.file_data = { filename: uploader.filename,
-                         doc_type: doc_type,
-                         description: description }
     end
 
     def uploader
